enable dependency injection of the ethercat master
By allowing the EtherCat master to be injected into each class, it is
possible to design a plugin-based architecture for binding the youbot
driver to a robotic software framework such as ROS or OROCOS: One
component or plugin in the framework instantiates the EtherCat master
and can then pass it to further components or plugins which are in turn
responsible for communicating with either the base, the manipulator or
the gripper.

The dependency injection is optional. Thus, the default behaviour remains unchanged.